Add files via upload
As a business analyst for Home Credit, you are supposed to first gather the information and clean it to make it usable.
The bureau information is at trade level, each individual trade level information is provided. You need to apply ‘Feature Engineering’ techniques to roll up the information at applicant level, and thereby create manual features for model building. 
Build a classification model to differentiate applicants between approves and rejects.

As a business analyst, you would want to find answers to the below questions for the bank:
How to leverage trade level information for Credit Bureaus by aggregating trade level information to applicant level in order to capture their payment behaviour?
Which application or payment behaviour factors significantly influence borrower’s behaviour on any new disbursed loan?
After identifying these factors, how to leverage them in the form of a model which can be used for decisioning?
Once the model is built, how to translate the model output into strategies and business insights for the bank?